Update config generation scripts
